requirements.txt
alembic==1.4.1
anytree==2.8.0
astroid==2.4.2
azure-core==1.10.0
azure-storage-blob==12.7.0
blis==0.7.4
boto3==1.16.57
botocore==1.19.57
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cryptography==3.3.1
cycler==0.10.0
cymem==2.0.5
databricks-cli==0.14.1
dataclasses==0.6
docker==4.4.1
entrypoints==0.3
filelock==3.0.12
fire==0.3.1
flake8==3.8.4
Flask==1.1.2
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
isodate==0.6.0
isort==5.6.4
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
kiwisolver==1.3.1
langcodes==2.1.0
lazy-object-proxy==1.4.3
Mako==1.1.4
marisa-trie==0.7.5
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mlflow==1.13.1
msgpack==1.0.0
msrest==0.6.19
murmurhash==1.0.5
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
OpenHowNet==0.0.1a11
packaging==20.4
pandas==1.1.4
Pillow==8.0.1
plac==1.1.3
plotly==4.14.3
preshed==3.0.5
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
protobuf==3.13.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-crf==0.7.0
pytorch-nlp==0.5.0
pytz==2020.4
pywsd==1.2.4
PyYAML==5.4
querystring-parser==1.2.4
regex==2020.10.23
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
revtok==0.0.3
s3transfer==0.3.4
sacremoses==0.0.43
scikit-learn==0.24.0
scipy==1.5.4
sentencepiece==0.1.94
six==1.15.0
sklearn==0.0
smmap==3.0.4
spacy==2.3.5
SQLAlchemy==1.3.22
sqlparse==0.4.1
srsly==1.0.5
tabulate==0.8.7
tensorboardX==2.1
termcolor==1.1.0
thinc==7.4.5
threadpoolctl==2.1.0
tokenizers==0.9.4
toml==0.10.2
#torch==1.7.1+cu110
#torchaudio==0.7.2
#torchtext==0.4.0
#torchvision==0.8.2+cu110
tqdm==4.56.0
transformers==4.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
wasabi==0.8.0
websocket-client==0.57.0
Werkzeug==1.0.1
wn==0.0.23
wordfreq==2.3.2
wrapt==1.12.1
zipp==3.4.0
pyinflect==0.5.1